Module Make.Transport

include Dns_client.S with type +'a io = 'a Lwt.t and type io_addr = [ `Plaintext of Ipaddr.t * int | `Tls of Tls.Config.client * Ipaddr.t * int ] with type stack = S.t * H.t

type context

A context is a network connection initialized by connect

type +'a io = 'a Lwt.t

io is the type of an effect. 'err is a polymorphic variant.

type io_addr = [

| `Plaintext of Ipaddr.t * int

| `Tls of Tls.Config.client * Ipaddr.t * int

]

An address for a given context type, usually this will consist of IP address + a TCP/IP or UDP/IP port number, but for some context types it can carry additional information for purposes of cryptographic verification.

type stack = S.t * H.t

A stack with which to connect.

type t

The abstract state of a DNS client.

val create : ?nameservers:(Dns.proto * io_addr list) -> timeout:int64 -> stack -> t

create ~nameservers ~timeout stack creates the state record of the DNS client. We use timeout (ns) as a cumulative time budget for connect and request timeouts.

val nameservers : t -> Dns.proto * io_addr list

The address of a nameservers that is supposed to work with the underlying context, can be used if the user does not want to bother with configuring their own.

val rng : int -> string

rng t is a random number generator.

val clock : unit -> int64

clock t is the monotonic clock.

val connect : t -> (Dns.proto * context, [> `Msg of string ]) Stdlib.result io

connect t is a new connection (context) to t, or an error.

val send_recv : context -> string -> (string, [> `Msg of string ]) Stdlib.result io

send_recv context buffer sends buffer to the context upstream, and then reads a buffer.

val close : context -> unit io

close context closes the context, freeing up resources.

val bind : 'a io -> ('a -> 'b io) -> 'b io

a.k.a. >>=

val lift : 'a -> 'a io

val happy_eyeballs : t -> happy_eyeballs